@@ -7,7 +7,10 @@ public class FakeListContributorsQueryService : IListContributorsQueryService
 {
   public Task<IEnumerable<ContributorDTO>> ListAsync()
   {
-    var result = new List<ContributorDTO>() { new ContributorDTO(1, "Ardalis"), new ContributorDTO(2, "Snowfrog") };
+    List<ContributorDTO> result = 
+        [new ContributorDTO(1, "Fake Contributor 1"), 
+        new ContributorDTO(2, "Fake Contributor 2")];
+
     return Task.FromResult(result.AsEnumerable());
   }
 }
